gui: refactoring
