+++ src/platform/windows/gui/debug/gpu.cpp
@@ -1,12 +1,19 @@
 #include "gpu.h"
+#include <fmt/core.h>
 #include <imgui.h>
+#include <magic_enum.hpp>
+#include <nlohmann/json.hpp>
+#include "platform/windows/gui/images.h"
+#include "renderer/opengl/opengl.h"
+#include "system.h"
+#include "utils/file.h"
 
 namespace gui::debug {
 
 void GPU::registersWindow(System *sys) {
     auto &gpu = sys->gpu;
 
-    ImGui::Begin("GPU", &showGpuWindow, ImVec2(200, 100));
+    ImGui::Begin("GPU", &registersWindowOpen, ImVec2(200, 100));
 
     int horRes = gpu->gp1_08.getHorizontalResoulution();
     int verRes = gpu->gp1_08.getVerticalResoulution();
@@ -31,11 +38,36 @@ void GPU::registersWindow(System *sys) {
     ImGui::End();
 }
 
+void replayCommands(gpu::GPU *gpu, int to) {
+    using gpu::Command;
+
+    auto commands = gpu->gpuLogList;
+    gpu->vram = gpu->prevVram;
+
+    gpu->gpuLogEnabled = false;
+    for (int i = 0; i <= to; i++) {
+        auto cmd = commands.at(i);
+
+        if (cmd.args.size() == 0) fmt::print("Panic! no args");
+
+        for (size_t j = 0; j < cmd.args.size(); j++) {
+            uint32_t arg = cmd.args[j];
+
+            if (j == 0) arg |= cmd.command << 24;
+
+            if (j == 0 && i == to && (cmd.cmd == Command::Polygon || cmd.cmd == Command::Rectangle || cmd.cmd == Command::Line)) {
+                arg = (cmd.command << 24) | 0x00ff00;
+                if (cmd.cmd == Command::Polygon || cmd.cmd == Command::Rectangle) arg &= ~(1 << 24);
+            }
+
+            gpu->write(0, arg);
+        }
+    }
+    gpu->gpuLogEnabled = true;
+}
+
 void GPU::logWindow(System *sys) {
     vramAreas.clear();
-    if (!gpuLogEnabled) {
-        return;
-    }
     static std::unique_ptr<Texture> textureImage;
     static std::vector<uint8_t> textureUnpacked;
     if (!textureImage) {
@@ -43,7 +75,7 @@ void GPU::logWindow(System *sys) {
         textureUnpacked.resize(512 * 512 * 4);
     }
 
-    ImGui::Begin("GPU Log", &gpuLogEnabled, ImVec2(300, 400));
+    ImGui::Begin("GPU Log", &logWindowOpen, ImVec2(300, 400));
 
     ImGui::BeginChild("GPU Log", ImVec2(0, -ImGui::GetItemsLineHeightWithSpacing()), false);
     ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(0, 0));
@@ -232,8 +264,79 @@ void GPU::logWindow(System *sys) {
     }
 }
 
+void GPU::vramWindow(gpu::GPU *gpu) {
+    static std::unique_ptr<Texture> vramImage;
+    static std::vector<uint8_t> vramUnpacked;
+    if (!vramImage) {
+        vramImage = std::make_unique<Texture>(gpu::VRAM_WIDTH, gpu::VRAM_HEIGHT, GL_RGBA, GL_RGBA, GL_UNSIGNED_BYTE, false);
+        vramUnpacked.resize(gpu::VRAM_WIDTH * gpu::VRAM_HEIGHT * 4);
+    }
+
+    // Update texture
+    for (int y = 0; y < gpu::VRAM_HEIGHT; y++) {
+        for (int x = 0; x < gpu::VRAM_WIDTH; x++) {
+            PSXColor c = gpu->vram[y * gpu::VRAM_WIDTH + x];
+
+            vramUnpacked[(y * gpu::VRAM_WIDTH + x) * 4 + 0] = c.r << 3;
+            vramUnpacked[(y * gpu::VRAM_WIDTH + x) * 4 + 1] = c.g << 3;
+            vramUnpacked[(y * gpu::VRAM_WIDTH + x) * 4 + 2] = c.b << 3;
+            vramUnpacked[(y * gpu::VRAM_WIDTH + x) * 4 + 3] = 0xff;
+        }
+    }
+    vramImage->update(vramUnpacked.data());
+
+    blinkTimer += 0.0025 * M_PI;
+    ImColor blinkColor = ImColor::HSV(blinkTimer, 1.f, 1.f, 0.75f);
+
+    ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0, 0));
+    ImGui::PushStyleColor(ImGuiCol_WindowBg, ImVec4(0.0, 0.0, 0.0, 1.0));
+
+    auto defaultSize = ImVec2(1024, 512 + ImGui::GetItemsLineHeightWithSpacing() * 2);
+    ImGui::SetNextWindowSizeConstraints(
+        ImVec2(defaultSize.x / 2, defaultSize.y / 2), ImVec2(defaultSize.x * 2, defaultSize.y * 2),
+        [](ImGuiSizeCallbackData *data) { data->DesiredSize.y = (data->DesiredSize.x / 2) + ImGui::GetItemsLineHeightWithSpacing() * 2; });
+
+    ImGui::Begin("VRAM", &vramWindowOpen, defaultSize, -1, ImGuiWindowFlags_NoScrollbar);
+
+    auto currentSize = ImGui::GetWindowContentRegionMax();
+    currentSize.y = currentSize.x / 2;
+
+    ImVec2 cursorPos = ImGui::GetCursorScreenPos();
+    ImGui::Image((ImTextureID)(uintptr_t)vramImage->get(), currentSize);
+
+    if (ImGui::Button("Original size")) {
+        ImGui::SetWindowSize(defaultSize);
+    }
+
+    ImGui::PopStyleColor();
+    ImGui::PopStyleVar();
+
+    float scale = currentSize.x / defaultSize.x;
+
+    ImDrawList *drawList = ImGui::GetWindowDrawList();
+    for (auto area : vramAreas) {
+        ImVec2 a, b;
+        a.x = cursorPos.x + area.pos.x * scale;
+        a.y = cursorPos.y + area.pos.y * scale;
+
+        b.x = cursorPos.x + (area.pos.x + area.size.x) * scale;
+        b.y = cursorPos.y + (area.pos.y + area.size.y) * scale;
+
+        drawList->AddRectFilled(a, b, blinkColor, 0.f, 0);
+
+        if (ImGui::IsMouseHoveringRect(a, b)) {
+            ImGui::BeginTooltip();
+            ImGui::TextUnformatted(area.name.c_str());
+            ImGui::EndTooltip();
+        }
+    }
+
+    ImGui::End();
+}
+
 void GPU::displayWindows(System *sys) {
-    if (registersWindowopen) registersWindow(sys);
+    if (registersWindowOpen) registersWindow(sys);
     if (logWindowOpen) logWindow(sys);
+    if (vramWindowOpen) vramWindow(sys->gpu.get());
 }
 };  // namespace gui::debug
