+++ src/platform/windows/gui/gui.h
@@ -8,9 +8,4 @@ void initGui();
 void deinitGui();
 void renderImgui(System* sys);
 extern bool showGui;
-extern bool gteRegistersEnabled;
-extern bool ioLogEnabled;
-extern bool gteLogEnabled;
-extern bool gpuLogEnabled;
 extern bool singleFrame;
-extern bool showIo;
